requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bson==0.5.10
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dnspython==2.3.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
Levenshtein==0.21.1
loguru==0.7.0
magic-filter==1.0.9
multidict==6.0.4
pip==21.3.1
psycopg2==2.9.6
pydantic==1.10.9
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
rapidfuzz==3.1.1
requests==2.31.0
setuptools==60.2.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
typing-extensions==4.7.0rc1
urllib3==2.0.3
wheel==0.37.1
win32-setctime==1.1.0
yarl==1.9.2
